creati 3 componenti per gestire le azioni del login, non utilizziamo più Login.js, sistemati CSS vari, aggiunte le 3 nuove Routes su App.js
