+++ src/components/LoginPswInstructions.css
@@ -1,0 +1,16 @@
+.Login-title-form {
+  text-align: center;
+  margin-bottom: 12px;
+  font-size: 21px;
+}
+.Login-text-form {
+  font-size: 17px;
+  font-weight: 300;
+  text-align: center;
+  margin-bottom: 40px;
+}
+@media (min-width: 768px) {
+  .Login-title-form {
+    font-size: 27px;
+  }
+}
